+++ score-module-application/src/main/java/ifood/score/domain/repository/entity/RelevanceCategoryMongo.java
@@ -3,8 +3,6 @@ package ifood.score.domain.repository.entity;
 import ifood.score.menu.Category;
 import lombok.*;
 
-import java.math.BigDecimal;
-
 @Setter
 @Getter
 @AllArgsConstructor
@@ -13,6 +11,6 @@ import java.math.BigDecimal;
 public class RelevanceCategoryMongo {
 
     private Category category;
-    private BigDecimal relevance;
+    private Double relevance;
 
 }
