+++ score-module-application/src/main/java/ifood/score/domain/repository/entity/ScoreCategoryMongo.java
@@ -9,8 +9,6 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
-import java.util.UUID;
 
 @Data
 @NoArgsConstructor
@@ -20,9 +18,10 @@ public class ScoreCategoryMongo implements Serializable {
 
     private static final long serialVersionUID = -839224893062958534L;
 
-    @Id @NonNull
+    @Id
+    @NonNull
     private Category category;
 
     @NonNull
-    private BigDecimal score;
+    private Double score;
 }
