+++ score-module-application/src/main/java/ifood/score/domain/repository/entity/ScoreMenuItemMongo.java
@@ -8,7 +8,6 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -19,9 +18,10 @@ public class ScoreMenuItemMongo implements Serializable {
 
     private static final long serialVersionUID = -839224893062958534L;
 
-    @Id @NonNull
+    @Id
+    @NonNull
     private UUID menuUuid;
 
     @NonNull
-    private BigDecimal score;
+    private Double score;
 }
